#5 feat: WelcomeView 뷰컨트롤러와 분리
