+++ 34th-Sopt-Seminar/Presentation/ViewControllers/WelcomeViewController_DelegatePattern.swift
@@ -15,91 +15,45 @@ protocol DataBindProtocol: AnyObject {
 final class WelcomeViewController_DelegatePattern: UIViewController {
     typealias handler = ((String) -> (Void))
     
+    private let rootView = WelcomeView()
+    private lazy var backToLoginButton = rootView.backToLoginButton
+    private lazy var welcomeLabel = rootView.welcomeLabel
     weak var delegate: DataBindProtocol?
-    var completionHandler: handler?
+//    var completionHandler: handler?
     var id: String?
-        
-    private let imageView: UIImageView = {
-       let imageView = UIImageView()
-        imageView.image = UIImage(named: "rabbit")
-        return imageView
-    }()
     
-    private let welcomeLabel: UILabel = {
-        let label = UILabel()
-        label.text = "???님\n반가워요!"
-        label.textColor = .black
-        label.font = .font(ofSize: 25, weight: .w800)
-        label.textAlignment = .center
-        label.numberOfLines = 2
-        return label
-    }()
-    
-    private var goHomeButton: UIButton = {
-        let button = UIButton()
-        button.backgroundColor = UIColor(red: 255/255, green: 111/255, blue: 15/255, alpha: 1)
-        button.setTitle("메인으로", for: .normal)
-        button.setTitleColor(.white, for: .normal)
-        button.titleLabel?.font = .font(ofSize: 18, weight: .w700)
-        button.layer.cornerRadius = 6
-        return button
-    }()
-    
-    private lazy var backToLoginButton: UIButton = {
-        let button = UIButton()
-        button.backgroundColor = UIColor(red: 221/255, green: 222/255, blue: 227/255, alpha: 1)
-        button.setTitle("다시 로그인", for: .normal)
-        button.setTitleColor(UIColor(red: 172/255, green: 176/255, blue: 185/255, alpha: 1), for: .normal)
-        button.titleLabel?.font = .font(ofSize: 18, weight: .w700)
-        button.addTarget(self, action: #selector(backToLoginButtonDidTap), for: .touchUpInside)
-        button.layer.cornerRadius = 6
-        return button
-    }()
+    override func loadView() {
+        super.loadView()
+        self.view = rootView
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         setStyle()
-        setHierarchy()
-        setLayout()
+        setButtonAction()
         bindID()
     }
+}
+
+extension WelcomeViewController_DelegatePattern {
     
-    private func setStyle() {
+    func setStyle() {
         self.navigationController?.isNavigationBarHidden = true
-        self.view.backgroundColor = .white
     }
     
-    private func setHierarchy() {
-        [imageView, welcomeLabel, goHomeButton, backToLoginButton].forEach {
-            self.view.addSubview($0)
-        }
+    func setButtonAction() {
+        backToLoginButton.addTarget(self, action: #selector(backToLoginButtonDidTap), for: .touchUpInside)
     }
     
-    private func setLayout() {
-        imageView.snp.makeConstraints {
-            $0.size.equalTo(150)
-            $0.top.equalTo(view.safeAreaLayoutGuide).offset(40)
-            $0.centerX.equalToSuperview()
-        }
+    private func bindID() {
+//        guard let idText = id else { return }
+//        welcomeLabel.text = "\(idText)님\n반가워요!"
         
-        welcomeLabel.snp.makeConstraints {
-            $0.top.equalTo(imageView.snp.bottom).offset(58)
-            $0.centerX.equalToSuperview()
-        }
-        
-        goHomeButton.snp.makeConstraints {
-            $0.height.equalTo(58)
-            $0.top.equalTo(welcomeLabel.snp.bottom).offset(71)
-            $0.horizontalEdges.equalToSuperview().inset(20)
-            $0.centerX.equalToSuperview()
-        }
-        
-        backToLoginButton.snp.makeConstraints {
-            $0.height.equalTo(58)
-            $0.top.equalTo(goHomeButton.snp.bottom).offset(14)
-            $0.horizontalEdges.equalToSuperview().inset(20)
-            $0.centerX.equalToSuperview()
+        if let idText = id {
+            welcomeLabel.text = "\(idText)님\n반가워요!"
+        } else {
+            print("id값이 없습니다.")
+            /// UITextField text는 nil값이 넘어올 일이 없다! textField가 비어있어도 nil이 아닌 ""가 반환됨. objective-c와의 호환성을 위해 optional 타입으로 지정한 것이라고 함.
         }
     }
     
@@ -116,17 +70,4 @@ final class WelcomeViewController_DelegatePattern: UIViewController {
             self.navigationController?.popViewController(animated: true)
         }
     }
-    
-    private func bindID() {
-//        guard let idText = id else { return }
-//        welcomeLabel.text = "\(idText)님\n반가워요!"
-        
-        if let idText = id {
-            welcomeLabel.text = "\(idText)님\n반가워요!"
-        } else {
-            print("id값이 없습니다.")
-            /// UITextField text는 nil값이 넘어올 일이 없다! textField가 비어있어도 nil이 아닌 ""가 반환됨. objective-c와의 호환성을 위해 optional 타입으로 지정한 것이라고 함.
-        }
-    }
 }
-
